Deal with empty cells when comparing query table results. Fixes #849.
